A helper function PointerOf to get pointer to anything
If you need a pointer to the literal value you need to declare
a variable first and than take an address of it. With the new
PointerOf function you don't need to declare a variable and
can use a call like `PointerOf(42)` in place to get a pointer
to `int` type with value `42`.